+++ server/controllers/TasksController.js
@@ -2,27 +2,19 @@ import BaseController from "../utils/BaseController";
 import Auth0Provider from "@bcwdev/auth0provider";
 import { tasksService } from "../services/TasksService";
 import { commentsService } from "../services/CommentsService"
+import { UnAuthorized } from "../utils/Errors";
 
 export class TasksController extends BaseController {
   constructor() {
     super("api/tasks");
     this.router
       .use(Auth0Provider.getAuthorizedUserInfo)
-      // .get("", this.getTasks)
-      // .get("/:taskId", this.getTask)
       .get("/:taskId/comments", this.getCommentsByTaskId)
       .post("", this.createTask)
       .put("", this.updateTask)
       .delete("/:taskId", this.deleteTask)
   }
 
-  // async getTask(req, res, next) {
-  //   try {
-
-  //   } catch (error) {
-  //     next(error);
-  //   }
-  // }
   async getCommentsByTaskId(req, res, next) {
     try {
       let comments = await commentsService.getCommentsByTaskId(req.params.taskId);
@@ -53,9 +45,9 @@ export class TasksController extends BaseController {
 
   async deleteTask(req, res, next) {
     try {
-      // if (req.body) TODO Check for creatorId
-      let task = await tasksService.deleteTask(req.params.taskId)
-      res.send(task);
+      if (req.body.creatorId != req.userInfo.id) throw new UnAuthorized("Unauthorized to delete this task");
+      await tasksService.deleteTask(req.params.taskId)
+      res.send("Task deleted!");
     } catch (error) {
       next(error);
     }
